move ShadowHostBuilder to "impl" package
